+++ Modules/SharePointDsc/DSCResources/MSFT_SPAccessServiceApp/MSFT_SPAccessServiceApp.psm1
@@ -123,7 +123,6 @@ function Set-TargetResource
                 $_.TypeName -eq "Access Services Web Service Application"  
             }
 
-            # Remove the connected proxy(ies)
             $proxies = Get-SPServiceApplicationProxy
             foreach($proxyInstance in $proxies)
             {
@@ -133,7 +132,7 @@ function Set-TargetResource
                 }
             }
 
-            Remove-SPServiceApplication $app -Confirm:$false
+            Remove-SPServiceApplication -Identity $app -Confirm:$false
         }
     }
 }