+++ Modules/SharePointDsc/DSCResources/MSFT_SPAppManagementServiceApp/MSFT_SPAppManagementServiceApp.psm1
@@ -198,7 +198,6 @@ function Set-TargetResource
                 $_.TypeName -eq "App Management Service Application"
             }
 
-            # Remove the connected proxy(ies)
             $proxies = Get-SPServiceApplicationProxy
             foreach($proxyInstance in $proxies)
             {
@@ -208,7 +207,7 @@ function Set-TargetResource
                 }
             }
 
-            Remove-SPServiceApplication $app -Confirm:$false
+            Remove-SPServiceApplication -Identity $app -Confirm:$false
         }
     }
 }