+++ Modules/SharePointDsc/DSCResources/MSFT_SPBCSServiceApp/MSFT_SPBCSServiceApp.psm1
@@ -159,7 +159,6 @@ function Set-TargetResource
                 $_.TypeName -eq "Business Data Connectivity Service Application"
             }
 
-            # Remove the connected proxy(ies)
             $proxies = Get-SPServiceApplicationProxy
             foreach($proxyInstance in $proxies)
             {
@@ -169,7 +168,7 @@ function Set-TargetResource
                 }
             }
 
-            Remove-SPServiceApplication $app -Confirm:$false
+            Remove-SPServiceApplication -Identity $app -Confirm:$false
         }
     }
 }