+++ Modules/SharePointDsc/DSCResources/MSFT_SPExcelServiceApp/MSFT_SPExcelServiceApp.psm1
@@ -128,7 +128,6 @@ function Set-TargetResource
                     $_.TypeName -eq "Excel Services Application Web Service Application"  
             }
 
-            # Remove the connected proxy(ies)
             $proxies = Get-SPServiceApplicationProxy
             foreach($proxyInstance in $proxies)
             {
@@ -138,7 +137,7 @@ function Set-TargetResource
                 }
             }
 
-            Remove-SPServiceApplication $app -Confirm:$false
+            Remove-SPServiceApplication -Identity $app -Confirm:$false
         }
     }
 }