+++ Tests/Unit/SharePointDsc/SharePointDsc.SPBCSServiceApp.Tests.ps1
@@ -32,7 +32,7 @@ Describe "SPBCSServiceApp - SharePoint Build $((Get-Item $SharePointCmdletModule
         Import-Module $Global:CurrentSharePointStubModule -WarningAction SilentlyContinue
 
         Mock Remove-SPServiceApplication { }
-        Mock Get-SPServiceApplicationProxy { return $null }
+        Mock Get-SPServiceApplicationProxy { }
 
         Context "When no service applications exist in the current farm and it should" {